requirements.txt
fastapi~=0.117.1
uvicorn
pymongo~=4.15.1
pydantic[email]~=2.11.9
python-multipart
passlib~=1.7.4
bcrypt
python-jose[cryptography]~=3.5.0
cloudinary~=1.44.1
boto3~=1.40.65
spacy~=3.8.7